requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
python-decouple>=3.8,<4.0
Pillow>=10.0,<11.0
mysqlclient>=2.2,<3.0   # MySQL database driver

# AI / LLM Providers — LLM Race Pattern (Gemini | DeepSeek | Claude | Groq)
openai>=1.0,<2.0              # DeepSeek uses OpenAI-compatible SDK
google-generativeai>=0.8,<1.0 # Google Gemini
anthropic>=0.30,<1.0          # Anthropic Claude
groq>=0.4,<1.0                # Groq (Llama 3)

# MCP — Model Context Protocol server + client SDK
mcp>=1.0,<2.0
duckduckgo-search>=8.0,<9.0
lxml>=6.0,<7.0

# Library Agent — RAG + Vector Store
chromadb>=1.0,<2.0
sentence-transformers>=3.0,<6.0
boto3>=1.40,<2.0              # Cloudflare R2 (S3-compatible)
django-storages>=1.14,<2.0    # Django R2/S3 storage backend

# Research Agent — Web scraping
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
pypdf>=4.0,<7.0               # PDF text extraction

# Async support
asgiref>=3.7,<4.0

# Testing
pytest>=9.0,<10.0
pytest-django>=4.9,<5.0
pytest-asyncio>=1.0,<2.0